Refactor and DRY up specs - method uses, account mappings and actions in before callbacks
