+++ spec/resource/account_creation_policy_spec.rb
@@ -28,13 +28,7 @@ describe Stormpath::Resource::AccountCreationPolicy, :vcr do
     end
 
     before do
-      test_api_client.account_store_mappings.create(
-        application: application,
-        account_store: directory,
-        list_index: 1,
-        is_default_account_store: false,
-        is_default_group_store: false
-      )
+      map_account_store(application, directory, 1, false, false)
     end
 
     after do
