+++ spec/resource/email_template_spec.rb
@@ -16,13 +16,7 @@ describe Stormpath::Resource::EmailTemplate, :vcr do
     let(:reset_email_template) { password_policy.reset_email_templates.first }
 
     before do
-      test_api_client.account_store_mappings.create(
-        application: application,
-        account_store: directory,
-        list_index: 1,
-        is_default_account_store: false,
-        is_default_group_store: false
-      )
+      map_account_store(application, directory, 1, false, false)
     end
 
     after { directory.delete }
