+++ spec/resource/password_strength_spec.rb
@@ -16,13 +16,7 @@ describe Stormpath::Resource::PasswordStrength, :vcr do
     let(:password_strength) { password_policy.strength }
 
     before do
-      test_api_client.account_store_mappings.create(
-        application: application,
-        account_store: directory,
-        list_index: 1,
-        is_default_account_store: false,
-        is_default_group_store: false
-      )
+      map_account_store(application, directory, 1, false, false)
     end
 
     after { directory.delete }
